[Refactor]헤더파일 분리
