+++ CRA_Project_Tester/testScript.h
@@ -1,0 +1,36 @@
+#pragma once
+#include "testShell.h"
+class SSDTest_FullWriteAndReadCompare :public ITestOperation
+{
+public:
+	SSDTest_FullWriteAndReadCompare(Write* w, Read* r) : write(w), read(r) {};
+	void run(string command1 = "", string command2 = "") override;
+private:
+	Write* write;
+	Read* read;
+};
+
+
+class SSDTest_PartialLBAWrite :public ITestOperation, public exception
+{
+public:
+	SSDTest_PartialLBAWrite(Write* w, Read* r) : mWrite(w), mRead(r) {};
+	SSDTest_PartialLBAWrite() : mWrite(nullptr), mRead(nullptr) {};
+	void run(string param1, string param2) override;
+
+private:
+	Write* mWrite;
+	Read* mRead;
+};
+
+class SSDTest_WriteReadAging :public ITestOperation, public exception
+{
+public:
+	SSDTest_WriteReadAging(Write* w, Read* r) : mWrite(w), mRead(r) {};
+	SSDTest_WriteReadAging() : mWrite(nullptr), mRead(nullptr) {};
+	void run(string param1, string param2) override;
+
+private:
+	Write* mWrite;
+	Read* mRead;
+};
